pyproject.toml
[project]
name = "tyagent"
version = "0.1.0"
description = "A lightweight, extensible agent framework. Prioritizes Feishu support."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "httpx>=0.27",
    "lark-oapi>=1.5.5",
    "pyyaml>=6.0",
    "qrcode>=8.2",
    "jieba>=0.42",
]

[project.optional-dependencies]
feishu = [
    "lark-oapi>=1.5",
    "qrcode>=7.4",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[project.scripts]
tyagent = "tyagent_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
